Refactor: AutoComplete Store 분리
